revert: restore .b class background color extraction for Tartrazine themes
The .b class IS the correct background color in Tartrazine themes.
Ascendancy/light correctly has #cccccc as its background color.

Previous commit incorrectly avoided .b class thinking it was variable styling,
but in Tartrazine themes, .b { background-color: #xxx } is always the editor
background color that should be used.

This reverts to the original correct logic that extracts background color from
the .b class and applies it to the #editor element.